backend/requirements.txt
Django
djangorestframework
channels
Faker
django-cors-headers
daphne
channels_redis
djangorestframework-simplejwt